combined databaseController classes
Combined the two databaseControllers we have for modularity
